@@ -1,43 +1,34 @@
 {
-  "name": "sportify",
-  "private": true,
-  "version": "0.0.0",
-  "type": "commonjs",
+  "name": "sportify-frontend",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "dist-electron/main.js",
   "scripts": {
     "dev": "vite",
-    "build": "tsc && vite build && electron-builder",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "build": "vite build",
     "preview": "vite preview"
   },
   "dependencies": {
+    "@tailwindcss/vite": "^4.1.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.5.0",
-    "react-router-dom": "^7.6.0",
-    "react-spinners": "^0.15.0",
-    "vite-tsconfig-paths": "^5.1.4"
+    "react-router-dom": "^6.30.0",
+    "sportify-frontend": "file:",
+    "tailwindcss": "^4.1.7"
   },
   "devDependencies": {
-    "@tailwindcss/postcss": "^4.1.6",
-    "@types/react": "^18.2.64",
-    "@types/react-dom": "^18.2.21",
-    "@typescript-eslint/eslint-plugin": "^7.1.1",
-    "@typescript-eslint/parser": "^7.1.1",
-    "@vitejs/plugin-react": "^4.2.1",
-    "autoprefixer": "^10.4.21",
-    "electron": "^30.0.1",
-    "electron-builder": "^24.13.3",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.5",
-    "less": "^4.3.0",
-    "postcss": "^8.5.3",
-    "tailwindcss": "^4.1.6",
-    "typescript": "^5.2.2",
-    "vite": "^5.1.6",
-    "vite-plugin-electron": "^0.28.6",
-    "vite-plugin-electron-renderer": "^0.14.5",
-    "vite-plugin-svgr": "^4.3.0"
-  },
-  "main": "dist-electron/main.js"
+    "@types/jest": "^29.5.14",
+    "@types/react": "^19.1.4",
+    "@types/react-dom": "^19.1.5",
+    "@vitejs/plugin-react": "^4.4.1",
+    "daisyui": "^5.0.35",
+    "electron": "^36.2.0",
+    "jest": "^29.7.0",
+    "vite": "^6.3.5",
+    "vite-plugin-electron": "^0.29.0",
+    "vite-plugin-electron-renderer": "^0.14.6",
+    "vite-plugin-svgr": "^4.3.0",
+    "vite-tsconfig-paths": "^5.1.4"
+  }
 }
