@@ -1,27 +1,33 @@
 import { defineConfig } from 'vite'
+import tailwindcss from '@tailwindcss/vite'
 import path from 'node:path'
 import electron from 'vite-plugin-electron/simple'
 import react from '@vitejs/plugin-react'
+import tsConfigPaths from 'vite-tsconfig-paths'
 import svgr from 'vite-plugin-svgr'
 
-// ✅ Use dynamic ESM-safe import for vite-tsconfig-paths
-export default defineConfig(async () => {
-  const { default: tsConfigPaths } = await import('vite-tsconfig-paths')
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [
+    tailwindcss(),
+    react(),
+    tsConfigPaths(),
+    svgr(),
+    electron({
+      main: {
+        // Shortcut of `build.lib.entry`.
+        entry: 'electron-main/main.ts',
+      },
+      preload: {
+        // Shortcut of `build.rollupOptions.input`.
+        // Preload scripts may contain Web assets, so use the `build.rollupOptions.input` instead `build.lib.entry`.
+        input: path.join(__dirname, 'electron-main/preload.ts'),
+      },
 
-  return {
-    plugins: [
-      react(),
-      tsConfigPaths(), // ✅ Now safe with ESM-only
-      svgr(),
-      electron({
-        main: {
-          entry: 'electron/main.ts',
-        },
-        preload: {
-          input: path.join(__dirname, 'electron/preload.ts'),
-        },
-        renderer: process.env.NODE_ENV === 'test' ? undefined : {},
-      }),
-    ],
-  }
+      renderer: process.env.NODE_ENV === 'test'
+        // https://github.com/electron-vite/vite-plugin-electron-renderer/issues/78#issuecomment-2053600808
+        ? undefined
+        : {},
+    }),
+  ],
 })
